♻️ refactor: 관심 스토리 훅함수 분리
