+++ apps/web/src/app/(needLogin)/mypage/like_story/_hooks/useLikeStory.ts
@@ -1,0 +1,12 @@
+import { useAtom } from 'jotai';
+
+import { likeStoriesAtom } from '@/store/mypage.atoms';
+
+export const useLikeStory = () => {
+  const [likeStories, setLikeStories] = useAtom(likeStoriesAtom);
+
+  return {
+    likeStories,
+    setLikeStories,
+  };
+};
