+++ apps/web/src/app/(needLogin)/mypage/like_story/_hooks/useViewStory.ts
@@ -1,0 +1,12 @@
+import { useAtom } from 'jotai';
+
+import { viewStoriesAtom } from '@/store/mypage.atoms';
+
+export const useViewStory = () => {
+  const [viewStories, setViewStories] = useAtom(viewStoriesAtom);
+
+  return {
+    viewStories,
+    setViewStories,
+  };
+};
